move cert wizard pages
