+++ web/src/pages/settings/SettingsCertificatesPage/SettingsCoreCertificateWizardPage/style.scss
