+++ web/src/pages/settings/SettingsCertificatesPage/SettingsEdgeCertificateWizardPage/steps/SettingsEdgeCertificateWizardExternalUrlSslConfigStep.tsx
@@ -1,19 +1,19 @@
 import { useQuery } from '@tanstack/react-query';
 import { useCallback, useEffect, useState } from 'react';
-import { m } from '../../../paraglide/messages';
-import api from '../../../shared/api/api';
-import { Controls } from '../../../shared/components/Controls/Controls';
-import { LoadingStep } from '../../../shared/components/LoadingStep/LoadingStep';
-import { WizardCard } from '../../../shared/components/wizard/WizardCard/WizardCard';
-import { Button } from '../../../shared/defguard-ui/components/Button/Button';
-import { CodeCard } from '../../../shared/defguard-ui/components/CodeCard/CodeCard';
-import { Divider } from '../../../shared/defguard-ui/components/Divider/Divider';
-import { SizedBox } from '../../../shared/defguard-ui/components/SizedBox/SizedBox';
-import { ThemeSpacing } from '../../../shared/defguard-ui/types';
-import { useSSEController } from '../../../shared/hooks/useSSEController';
-import { downloadFile } from '../../../shared/utils/download';
-import caIcon from '../../SetupPage/assets/ca.png';
-import '../../SetupPage/autoAdoption/steps/style.scss';
+import { m } from '../../../../../paraglide/messages';
+import api from '../../../../../shared/api/api';
+import { Controls } from '../../../../../shared/components/Controls/Controls';
+import { LoadingStep } from '../../../../../shared/components/LoadingStep/LoadingStep';
+import { WizardCard } from '../../../../../shared/components/wizard/WizardCard/WizardCard';
+import { Button } from '../../../../../shared/defguard-ui/components/Button/Button';
+import { CodeCard } from '../../../../../shared/defguard-ui/components/CodeCard/CodeCard';
+import { Divider } from '../../../../../shared/defguard-ui/components/Divider/Divider';
+import { SizedBox } from '../../../../../shared/defguard-ui/components/SizedBox/SizedBox';
+import { ThemeSpacing } from '../../../../../shared/defguard-ui/types';
+import { useSSEController } from '../../../../../shared/hooks/useSSEController';
+import { downloadFile } from '../../../../../shared/utils/download';
+import caIcon from '../../../../SetupPage/assets/ca.png';
+import '../../../../SetupPage/autoAdoption/steps/style.scss';
 import { SettingsEdgeCertificateWizardStep } from '../types';
 import { useSettingsEdgeCertificateWizardStore } from '../useSettingsEdgeCertificateWizardStore';
