+++ web/src/pages/settings/SettingsCertificatesPage/SettingsEdgeCertificateWizardPage/steps/SettingsEdgeCertificateWizardSummaryStep.tsx
@@ -1,11 +1,11 @@
 import { useQueryClient } from '@tanstack/react-query';
 import { useNavigate } from '@tanstack/react-router';
-import { m } from '../../../paraglide/messages';
-import { Controls } from '../../../shared/components/Controls/Controls';
-import { WizardCard } from '../../../shared/components/wizard/WizardCard/WizardCard';
-import { Button } from '../../../shared/defguard-ui/components/Button/Button';
-import { SizedBox } from '../../../shared/defguard-ui/components/SizedBox/SizedBox';
-import { ThemeSpacing } from '../../../shared/defguard-ui/types';
+import { m } from '../../../../../paraglide/messages';
+import { Controls } from '../../../../../shared/components/Controls/Controls';
+import { WizardCard } from '../../../../../shared/components/wizard/WizardCard/WizardCard';
+import { Button } from '../../../../../shared/defguard-ui/components/Button/Button';
+import { SizedBox } from '../../../../../shared/defguard-ui/components/SizedBox/SizedBox';
+import { ThemeSpacing } from '../../../../../shared/defguard-ui/types';
 import { useSettingsEdgeCertificateWizardStore } from '../useSettingsEdgeCertificateWizardStore';
 import '../style.scss';
