+++ web/src/pages/settings/SettingsCertificatesPage/SettingsEdgeCertificateWizardPage/style.scss
