+++ web/src/routes/_authorized/_wizard/settings-core-certificate.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router';
-import { SettingsCoreCertificateWizardPage } from '../../../pages/SettingsCoreCertificateWizardPage/SettingsCoreCertificateWizardPage';
+import { SettingsCoreCertificateWizardPage } from '../../../pages/settings/SettingsCertificatesPage/SettingsCoreCertificateWizardPage/SettingsCoreCertificateWizardPage';
 
 export const Route = createFileRoute('/_authorized/_wizard/settings-core-certificate')({
   component: SettingsCoreCertificateWizardPage,
