+++ web/src/routes/_authorized/_wizard/settings-edge-certificate.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router';
-import { SettingsEdgeCertificateWizardPage } from '../../../pages/SettingsEdgeCertificateWizardPage/SettingsEdgeCertificateWizardPage';
+import { SettingsEdgeCertificateWizardPage } from '../../../pages/settings/SettingsCertificatesPage/SettingsEdgeCertificateWizardPage/SettingsEdgeCertificateWizardPage';
 
 export const Route = createFileRoute('/_authorized/_wizard/settings-edge-certificate')({
   component: SettingsEdgeCertificateWizardPage,
